+++ core/app/identity.go
@@ -1,3 +1,4 @@
 package app
 
 var RuntimeIdentity = []byte("runtime")
+var ConsensusIdentity = []byte("consensus")