+++ runtime/v2/module.go
@@ -4,6 +4,9 @@ import (
 	"fmt"
 	"os"
 
+	"cosmossdk.io/core/genesis"
+	"github.com/cosmos/cosmos-sdk/baseapp"
+
 	"github.com/cosmos/gogoproto/proto"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/reflect/protodesc"
@@ -25,6 +28,7 @@ import (
 	"cosmossdk.io/runtime/v2/services"
 	"cosmossdk.io/server/v2/stf"
 	storetypes "cosmossdk.io/store/types"
+	rootstorev2 "cosmossdk.io/store/v2/root"
 	"cosmossdk.io/x/tx/signing"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -100,19 +104,20 @@ func (m appModule) AutoCLIOptions() *autocliv1.ModuleOptions {
 func init() {
 	appconfig.Register(&runtimev2.Module{},
 		appconfig.Provide(
-			ProvideApp,
+			ProvideAppBuilder,
 			ProvideInterfaceRegistry,
 			ProvideKVStoreKey,
 			ProvideEnvironment,
 			ProvideModuleManager,
 			ProvideMemoryStoreKey,
 			ProvideAddressCodec,
+			ProvideGenesisTxHandler,
+			ProvideAppVersionModifier,
 		),
-		appconfig.Invoke(SetupAppBuilder),
 	)
 }
 
-func ProvideApp(interfaceRegistry codectypes.InterfaceRegistry) (
+func ProvideAppBuilder(interfaceRegistry codectypes.InterfaceRegistry) (
 	codec.Codec,
 	*codec.LegacyAmino,
 	*AppBuilder,
@@ -120,7 +125,6 @@ func ProvideApp(interfaceRegistry codectypes.InterfaceRegistry) (
 	appmodulev2.AppModule,
 	protodesc.Resolver,
 	protoregistry.MessageTypeResolver,
-	error,
 ) {
 	protoFiles := proto.HybridResolver
 	protoTypes := protoregistry.GlobalTypes
@@ -140,15 +144,16 @@ func ProvideApp(interfaceRegistry codectypes.InterfaceRegistry) (
 	cdc := codec.NewProtoCodec(interfaceRegistry)
 	msgRouterBuilder := stf.NewMsgRouterBuilder()
 	app := &App{
-		storeKeys:         nil,
-		interfaceRegistry: interfaceRegistry,
-		cdc:               cdc,
-		amino:             amino,
-		msgRouterBuilder:  msgRouterBuilder,
+		storeKeys:          nil,
+		interfaceRegistry:  interfaceRegistry,
+		cdc:                cdc,
+		amino:              amino,
+		msgRouterBuilder:   msgRouterBuilder,
+		queryRouterBuilder: stf.NewMsgRouterBuilder(), // TODO dedicated query router
 	}
 	appBuilder := &AppBuilder{app: app}
 
-	return cdc, amino, appBuilder, msgRouterBuilder, appModule{app}, protoFiles, protoTypes, nil
+	return cdc, amino, appBuilder, msgRouterBuilder, appModule{app}, protoFiles, protoTypes
 }
 
 type AppInputs struct {
@@ -161,6 +166,7 @@ type AppInputs struct {
 	InterfaceRegistry codectypes.InterfaceRegistry
 	LegacyAmino       *codec.LegacyAmino
 	Logger            log.Logger
+	StoreOptions      *rootstorev2.FactoryOptions
 }
 
 func SetupAppBuilder(inputs AppInputs) {
@@ -171,13 +177,29 @@ func SetupAppBuilder(inputs AppInputs) {
 	app.moduleManager = inputs.ModuleManager
 	app.moduleManager.RegisterInterfaces(inputs.InterfaceRegistry)
 	app.moduleManager.RegisterLegacyAminoCodec(inputs.LegacyAmino)
+
+	// TODO: this is a bit of a hack, but it's the only way to get the store keys into the app
+	// registerStoreKey could instead set this on StoreOptions directly
+	inputs.AppBuilder.storeOptions = inputs.StoreOptions
+	for _, sk := range inputs.AppBuilder.app.storeKeys {
+		inputs.AppBuilder.storeOptions.StoreKeys = append(inputs.AppBuilder.storeOptions.StoreKeys, sk.String())
+	}
 }
 
-func ProvideModuleManager(logger log.Logger, cdc codec.Codec, config *runtimev2.Module, modules map[string]appmodulev2.AppModule) *MM {
+func ProvideModuleManager(
+	logger log.Logger,
+	cdc codec.Codec,
+	config *runtimev2.Module,
+	modules map[string]appmodulev2.AppModule,
+) *MM {
 	return NewModuleManager(logger, cdc, config, modules)
 }
 
-func ProvideInterfaceRegistry(addressCodec address.Codec, validatorAddressCodec ValidatorAddressCodec, customGetSigners []signing.CustomGetSigner) (codectypes.InterfaceRegistry, error) {
+func ProvideInterfaceRegistry(
+	addressCodec address.Codec,
+	validatorAddressCodec address.ValidatorAddressCodec,
+	customGetSigners []signing.CustomGetSigner,
+) (codectypes.InterfaceRegistry, error) {
 	signingOptions := signing.Options{
 		AddressCodec:          addressCodec,
 		ValidatorAddressCodec: validatorAddressCodec,
@@ -252,7 +274,7 @@ func ProvideEnvironment(logger log.Logger, config *runtimev2.Module, key depinje
 		BranchService:   nil, // TODO
 		EventService:    stf.NewEventService(),
 		GasService:      stf.NewGasMeterService(),
-		HeaderService:   nil, // TODO
+		HeaderService:   stf.HeaderService{},
 		KVStoreService:  kvService,
 		MemStoreService: memService,
 	}
@@ -260,28 +282,20 @@ func ProvideEnvironment(logger log.Logger, config *runtimev2.Module, key depinje
 	return env, kvService, memService
 }
 
-type (
-	// ValidatorAddressCodec is an alias for address.Codec for validator addresses.
-	ValidatorAddressCodec address.Codec
-
-	// ConsensusAddressCodec is an alias for address.Codec for validator consensus addresses.
-	ConsensusAddressCodec address.Codec
-)
-
 type AddressCodecInputs struct {
 	depinject.In
 
 	AuthConfig    *authmodulev1.Module    `optional:"true"`
 	StakingConfig *stakingmodulev1.Module `optional:"true"`
 
-	AddressCodecFactory          func() address.Codec         `optional:"true"`
-	ValidatorAddressCodecFactory func() ValidatorAddressCodec `optional:"true"`
-	ConsensusAddressCodecFactory func() ConsensusAddressCodec `optional:"true"`
+	AddressCodecFactory          func() address.Codec                 `optional:"true"`
+	ValidatorAddressCodecFactory func() address.ValidatorAddressCodec `optional:"true"`
+	ConsensusAddressCodecFactory func() address.ConsensusAddressCodec `optional:"true"`
 }
 
 // ProvideAddressCodec provides an address.Codec to the container for any
 // modules that want to do address string <> bytes conversion.
-func ProvideAddressCodec(in AddressCodecInputs) (address.Codec, ValidatorAddressCodec, ConsensusAddressCodec) {
+func ProvideAddressCodec(in AddressCodecInputs) (address.Codec, address.ValidatorAddressCodec, address.ConsensusAddressCodec) {
 	if in.AddressCodecFactory != nil && in.ValidatorAddressCodecFactory != nil && in.ConsensusAddressCodecFactory != nil {
 		return in.AddressCodecFactory(), in.ValidatorAddressCodecFactory(), in.ConsensusAddressCodecFactory()
 	}
@@ -305,4 +319,12 @@ func ProvideAddressCodec(in AddressCodecInputs) (address.Codec, ValidatorAddress
 	return addresscodec.NewBech32Codec(in.AuthConfig.Bech32Prefix),
 		addresscodec.NewBech32Codec(in.StakingConfig.Bech32PrefixValidator),
 		addresscodec.NewBech32Codec(in.StakingConfig.Bech32PrefixConsensus)
+}
+
+func ProvideGenesisTxHandler(appBuilder *AppBuilder) genesis.TxHandler {
+	return appBuilder.app
+}
+
+func ProvideAppVersionModifier(app *AppBuilder) baseapp.AppVersionModifier {
+	return app.app
 }